% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{getStratumPolygonFromGeojsonfFile}
\alias{getStratumPolygonFromGeojsonfFile}
\title{getStratumPolygonFromGeojsonfFile}
\usage{
getStratumPolygonFromGeojsonfFile(geojsonFilePath, textFilePath = NULL)
}
\arguments{
\item{geojsonFilePath}{path to the geojsonfile}
}
\value{
out a data.table object of the stratum definition
}
\description{
This function grap the stratum definition
}
