% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_AgeLength}
\alias{plot_AgeLength}
\title{plot_AgeLength}
\usage{
plot_AgeLength(projectPath, doc)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
doc a doc object to store figures to word document
}
\description{
This function plot the age vs length information in biotic
}
