% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_Baseline_processes}
\alias{plot_Baseline_processes}
\title{plot_Baseline_processes}
\usage{
plot_Baseline_processes(projectPath, doc)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
doc a doc object to store figures to word document
}
\description{
This function show the how the processes are connected
}
