% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_BeamKey}
\alias{plot_BeamKey}
\title{plot_progression}
\usage{
plot_BeamKey(projectPath, doc)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
doc a doc object to store figures to word document
}
\description{
This function plot the progression of vessels from a Stox project
}
