% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_Imputed_link}
\alias{plot_Imputed_link}
\title{plot_Imputed_link}
\usage{
plot_Imputed_link(projectPath, doc = NULL)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
A ggplot figure
}
\description{
This figure show the linkege where the impute has revieced samples from
}
