% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_SampleNumber}
\alias{plot_SampleNumber}
\title{plot_SampleNumber}
\usage{
plot_SampleNumber(projectPath, doc = NULL)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
A ggplot figure
}
\description{
This function show the sample number
}
