% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AutoFigures.R
\name{plot_StationLink}
\alias{plot_StationLink}
\title{plot_StationLink}
\usage{
plot_StationLink(projectPath, doc = NULL)
}
\arguments{
\item{projectPath}{Path to the stox project}

\item{doc}{a doc object to store figures to word document}
}
\value{
A ggplot figure
}
\description{
This function reads the stox project and grap the MeanNASCdata object
A figure is then made to view the mean NASC values
}
